+++ src/components/2018/page-pwa-with-ionic-angular-wordpress-api-5/page-pwa-with-ionic-angular-wordpress-api-5.tsx
@@ -515,6 +515,33 @@ return this.page < this.totalPages;
 
                         </ion-col>
                         <ion-col size-xs="12" size-sm="12" size-md="4" size-lg="4" size-xl="5">
+                            <ion-card>
+                                <ion-card-header>
+                                    <ion-card-subtitle>Build a PWA with Ionic, Angular, and the WordPress REST API</ion-card-subtitle>
+                                    <ion-card-title>Series content</ion-card-title>
+                                </ion-card-header>
+                                <ion-card-content>
+                                    <ion-list>
+                                        <ion-item href="/pwa-with-ionic-angular-wordpress-api-1">
+                                            <ion-label>Part 1</ion-label>
+                                        </ion-item>
+                                        <ion-item href="/pwa-with-ionic-angular-wordpress-api-2">
+                                            <ion-label>Part 2</ion-label>
+                                        </ion-item>
+                                        <ion-item href="/pwa-with-ionic-angular-wordpress-api-3">
+                                            <ion-label>Part 3</ion-label>
+                                        </ion-item>
+                                        <ion-item href="/pwa-with-ionic-angular-wordpress-api-4">
+                                            <ion-label>Part 4</ion-label>
+                                        </ion-item>
+                                        <ion-item href="/pwa-with-ionic-angular-wordpress-api-5">
+                                            <ion-icon name="chevron-forward-outline" slot="start"></ion-icon>
+                                            <ion-label>Part 5</ion-label>
+                                        </ion-item>
+                                    </ion-list>
+                                </ion-card-content>
+                            </ion-card>
+
                             <gls-adsense-ad />
                         </ion-col>
                     </ion-row>